Move evmread data in tests under evm package to separate happy and negative scenarios
